+++ docs_src/src/changelog.md
@@ -2,6 +2,10 @@
 
 ## V0.6.6
 
+- Updates to support the Yulan DLC:
+  - Fixed reading of Yulan troop upgrade costs and building upgrade choices
+  - Add label for Yulan maps on the lobby screen
+  - Fixed narration of leap ability
 - Text on story screens are now buttons, allowing `Enter` to advance to the next screen. Fixes having to tab to the next button to advance to the next screen
 - Improved Portuguese translations for generated essence and hostile troop labels